+++ commands/Help.js
@@ -2,6 +2,7 @@ const Command = require('./Command.js');
 const fs = require('fs');
 
 const categories = ["Codes", "Fun", "Technical", "Misc."];
+const color = 0x86D0CF;
 
 // Example usage of command: r!help 8ball
 // Potential usage of command: r!help r!8ball
@@ -33,7 +34,16 @@ class Help extends Command {
         for (var i = 0; i < obj.length; i++) {
           for (var j = 0; j < obj[i].length; j++) {
             if (obj[i][j].name.toLowerCase() == parsable.toLowerCase()) {
-              msg.channel.send("**r!" + obj[i][j].name + "**\n" + obj[i][j].help);
+              msg.channel.send({
+                embed: {
+                  author: {
+                    name: "Help"
+                  },
+                  title: "**r!" + obj[i][j].name + "**",
+                  color: color,
+                  description: obj[i][j].help,
+                }
+              });
               found = true;
               break;
             }
@@ -47,14 +57,40 @@ class Help extends Command {
       fs.readFile('./commands/commands.json', 'utf8', function(err, data) {
         if (err) throw err;
         var obj = JSON.parse(data);
-        var message = "**Command List**\nBasic command structure is `r![command]`. Commands are *not* case-sensitive. Use `r!help [command]` for more information about that command.\n";
+        var message = [];
         for (var i = 0; i < obj.length; i++) {
-          message += "\n**" + categories[i] + "** - ";
+          message[i] = "";
           for (var j = 0; j < obj[i].length; j++) {
-            message += "`" + obj[i][j].name + "` ";
+            message[i] += "`" + obj[i][j].name + "` ";
           }
         }
-        msg.channel.send(message);
+        msg.channel.send({
+          embed: {
+            author: {
+              name: "Help"
+            },
+            title: "**Command List**",
+            color: color,
+            description: "Basic command structure is `r![command]`. Commands are *not* case-sensitive. Use `r!help [command]` for more information about that command.",
+            fields: [{
+                name: categories[0],
+                value: message[0]
+              },
+              {
+                name: categories[1],
+                value: message[1]
+              },
+              {
+                name: categories[2],
+                value: message[2]
+              },
+              {
+                name: categories[3],
+                value: message[3]
+              }
+            ]
+          }
+        });
       });
     }
   }